Cleaned up Driver.kt so that only the relevant bits are exposed as public API
