+++ testing/node-driver/src/main/kotlin/net/corda/testing/internal/InternalTestUtils.kt
@@ -1,0 +1,93 @@
+package net.corda.testing.internal
+
+import net.corda.core.CordaException
+import net.corda.core.concurrent.CordaFuture
+import net.corda.core.internal.concurrent.openFuture
+import net.corda.core.internal.times
+import net.corda.core.utilities.NetworkHostAndPort
+import net.corda.core.utilities.getOrThrow
+import net.corda.core.utilities.millis
+import net.corda.core.utilities.seconds
+import org.slf4j.LoggerFactory
+import java.net.Socket
+import java.net.SocketException
+import java.time.Duration
+import java.util.concurrent.ScheduledExecutorService
+import java.util.concurrent.TimeUnit
+
+private val log = LoggerFactory.getLogger("net.corda.testing.internal.InternalTestUtils")
+
+/**
+ * @throws ListenProcessDeathException if [listenProcess] dies before the check succeeds, i.e. the check can't succeed as intended.
+ */
+fun addressMustBeBound(executorService: ScheduledExecutorService, hostAndPort: NetworkHostAndPort, listenProcess: Process? = null) {
+    addressMustBeBoundFuture(executorService, hostAndPort, listenProcess).getOrThrow()
+}
+
+fun addressMustBeBoundFuture(executorService: ScheduledExecutorService, hostAndPort: NetworkHostAndPort, listenProcess: Process? = null): CordaFuture<Unit> {
+    return poll(executorService, "address $hostAndPort to bind") {
+        if (listenProcess != null && !listenProcess.isAlive) {
+            throw ListenProcessDeathException(hostAndPort, listenProcess)
+        }
+        try {
+            Socket(hostAndPort.host, hostAndPort.port).close()
+            Unit
+        } catch (_exception: SocketException) {
+            null
+        }
+    }
+}
+
+/*
+ * The default timeout value of 40 seconds have been chosen based on previous node shutdown time estimate.
+ * It's been observed that nodes can take up to 30 seconds to shut down, so just to stay on the safe side the 60 seconds
+ * timeout has been chosen.
+ */
+fun addressMustNotBeBound(executorService: ScheduledExecutorService, hostAndPort: NetworkHostAndPort, timeout: Duration = 40.seconds) {
+    addressMustNotBeBoundFuture(executorService, hostAndPort).getOrThrow(timeout)
+}
+
+fun addressMustNotBeBoundFuture(executorService: ScheduledExecutorService, hostAndPort: NetworkHostAndPort): CordaFuture<Unit> {
+    return poll(executorService, "address $hostAndPort to unbind") {
+        try {
+            Socket(hostAndPort.host, hostAndPort.port).close()
+            null
+        } catch (_exception: SocketException) {
+            Unit
+        }
+    }
+}
+
+fun <A> poll(
+        executorService: ScheduledExecutorService,
+        pollName: String,
+        pollInterval: Duration = 500.millis,
+        warnCount: Int = 120,
+        check: () -> A?
+): CordaFuture<A> {
+    val resultFuture = openFuture<A>()
+    val task = object : Runnable {
+        var counter = -1
+        override fun run() {
+            if (resultFuture.isCancelled) return // Give up, caller can no longer get the result.
+            if (++counter == warnCount) {
+                log.warn("Been polling $pollName for ${(pollInterval * warnCount.toLong()).seconds} seconds...")
+            }
+            try {
+                val checkResult = check()
+                if (checkResult != null) {
+                    resultFuture.set(checkResult)
+                } else {
+                    executorService.schedule(this, pollInterval.toMillis(), TimeUnit.MILLISECONDS)
+                }
+            } catch (t: Throwable) {
+                resultFuture.setException(t)
+            }
+        }
+    }
+    executorService.submit(task) // The check may be expensive, so always run it in the background even the first time.
+    return resultFuture
+}
+
+class ListenProcessDeathException(hostAndPort: NetworkHostAndPort, listenProcess: Process) :
+        CordaException("The process that was expected to listen on $hostAndPort has died with status: ${listenProcess.exitValue()}")
