+++ testing/node-driver/src/main/kotlin/net/corda/testing/internal/RPCDriver.kt
@@ -52,158 +52,24 @@ import java.nio.file.Path
 import java.nio.file.Paths
 import java.util.*
 
-interface RPCDriverExposedDSLInterface : DriverDSLExposedInterface {
-    /**
-     * Starts an In-VM RPC server. Note that only a single one may be started.
-     *
-     * @param rpcUser The single user who can access the server through RPC, and their permissions.
-     * @param nodeLegalName The legal name of the node to check against to authenticate a super user.
-     * @param configuration The RPC server configuration.
-     * @param ops The server-side implementation of the RPC interface.
-     */
-    fun <I : RPCOps> startInVmRpcServer(
-            rpcUser: User = rpcTestUser,
-            nodeLegalName: CordaX500Name = fakeNodeLegalName,
-            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
-            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
-            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
-            ops: I
-    ): CordaFuture<RpcServerHandle>
-
-    /**
-     * Starts an In-VM RPC client.
-     *
-     * @param rpcOpsClass The [Class] of the RPC interface.
-     * @param username The username to authenticate with.
-     * @param password The password to authenticate with.
-     * @param configuration The RPC client configuration.
-     */
-    fun <I : RPCOps> startInVmRpcClient(
-            rpcOpsClass: Class<I>,
-            username: String = rpcTestUser.username,
-            password: String = rpcTestUser.password,
-            configuration: RPCClientConfiguration = RPCClientConfiguration.default
-    ): CordaFuture<I>
-
-    /**
-     * Starts an In-VM Artemis session connecting to the RPC server.
-     *
-     * @param username The username to authenticate with.
-     * @param password The password to authenticate with.
-     */
-    fun startInVmArtemisSession(
-            username: String = rpcTestUser.username,
-            password: String = rpcTestUser.password
-    ): ClientSession
-
-    /**
-     * Starts a Netty RPC server.
-     *
-     * @param serverName The name of the server, to be used for the folder created for Artemis files.
-     * @param rpcUser The single user who can access the server through RPC, and their permissions.
-     * @param nodeLegalName The legal name of the node to check against to authenticate a super user.
-     * @param configuration The RPC server configuration.
-     * @param ops The server-side implementation of the RPC interface.
-     */
-    fun <I : RPCOps> startRpcServer(
-            serverName: String = "driver-rpc-server-${random63BitValue()}",
-            rpcUser: User = rpcTestUser,
-            nodeLegalName: CordaX500Name = fakeNodeLegalName,
-            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
-            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
-            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
-            customPort: NetworkHostAndPort? = null,
-            ops: I
-    ): CordaFuture<RpcServerHandle>
-
-    /**
-     * Starts a Netty RPC client.
-     *
-     * @param rpcOpsClass The [Class] of the RPC interface.
-     * @param rpcAddress The address of the RPC server to connect to.
-     * @param username The username to authenticate with.
-     * @param password The password to authenticate with.
-     * @param configuration The RPC client configuration.
-     */
-    fun <I : RPCOps> startRpcClient(
-            rpcOpsClass: Class<I>,
-            rpcAddress: NetworkHostAndPort,
-            username: String = rpcTestUser.username,
-            password: String = rpcTestUser.password,
-            configuration: RPCClientConfiguration = RPCClientConfiguration.default
-    ): CordaFuture<I>
-
-    /**
-     * Starts a Netty RPC client in a new JVM process that calls random RPCs with random arguments.
-     *
-     * @param rpcOpsClass The [Class] of the RPC interface.
-     * @param rpcAddress The address of the RPC server to connect to.
-     * @param username The username to authenticate with.
-     * @param password The password to authenticate with.
-     */
-    fun <I : RPCOps> startRandomRpcClient(
-            rpcOpsClass: Class<I>,
-            rpcAddress: NetworkHostAndPort,
-            username: String = rpcTestUser.username,
-            password: String = rpcTestUser.password
-    ): CordaFuture<Process>
-
-    /**
-     * Starts a Netty Artemis session connecting to an RPC server.
-     *
-     * @param rpcAddress The address of the RPC server.
-     * @param username The username to authenticate with.
-     * @param password The password to authenticate with.
-     */
-    fun startArtemisSession(
-            rpcAddress: NetworkHostAndPort,
-            username: String = rpcTestUser.username,
-            password: String = rpcTestUser.password
-    ): ClientSession
-
-    fun startRpcBroker(
-            serverName: String = "driver-rpc-server-${random63BitValue()}",
-            rpcUser: User = rpcTestUser,
-            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
-            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
-            customPort: NetworkHostAndPort? = null
-    ): CordaFuture<RpcBrokerHandle>
-
-    fun startInVmRpcBroker(
-            rpcUser: User = rpcTestUser,
-            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
-            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE
-    ): CordaFuture<RpcBrokerHandle>
-
-    fun <I : RPCOps> startRpcServerWithBrokerRunning(
-            rpcUser: User = rpcTestUser,
-            nodeLegalName: CordaX500Name = fakeNodeLegalName,
-            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
-            ops: I,
-            brokerHandle: RpcBrokerHandle
-    ): RpcServerHandle
-}
-
-inline fun <reified I : RPCOps> RPCDriverExposedDSLInterface.startInVmRpcClient(
+inline fun <reified I : RPCOps> RPCDriverDSL.startInVmRpcClient(
         username: String = rpcTestUser.username,
         password: String = rpcTestUser.password,
         configuration: RPCClientConfiguration = RPCClientConfiguration.default
 ) = startInVmRpcClient(I::class.java, username, password, configuration)
 
-inline fun <reified I : RPCOps> RPCDriverExposedDSLInterface.startRandomRpcClient(
+inline fun <reified I : RPCOps> RPCDriverDSL.startRandomRpcClient(
         hostAndPort: NetworkHostAndPort,
         username: String = rpcTestUser.username,
         password: String = rpcTestUser.password
 ) = startRandomRpcClient(I::class.java, hostAndPort, username, password)
 
-inline fun <reified I : RPCOps> RPCDriverExposedDSLInterface.startRpcClient(
+inline fun <reified I : RPCOps> RPCDriverDSL.startRpcClient(
         rpcAddress: NetworkHostAndPort,
         username: String = rpcTestUser.username,
         password: String = rpcTestUser.password,
         configuration: RPCClientConfiguration = RPCClientConfiguration.default
 ) = startRpcClient(I::class.java, rpcAddress, username, password, configuration)
-
-interface RPCDriverInternalDSLInterface : DriverDSLInternalInterface, RPCDriverExposedDSLInterface
 
 data class RpcBrokerHandle(
         val hostAndPort: NetworkHostAndPort?,
@@ -235,26 +101,28 @@ fun <A> rpcDriver(
         extraCordappPackagesToScan: List<String> = emptyList(),
         notarySpecs: List<NotarySpec> = emptyList(),
         externalTrace: Trace? = null,
-        dsl: RPCDriverExposedDSLInterface.() -> A
-) = genericDriver(
-        driverDsl = RPCDriverDSL(
-                DriverDSL(
-                        portAllocation = portAllocation,
-                        debugPortAllocation = debugPortAllocation,
-                        systemProperties = systemProperties,
-                        driverDirectory = driverDirectory.toAbsolutePath(),
-                        useTestClock = useTestClock,
-                        isDebug = isDebug,
-                        startNodesInProcess = startNodesInProcess,
-                        waitForNodesToFinish = waitForNodesToFinish,
-                        extraCordappPackagesToScan = extraCordappPackagesToScan,
-                        notarySpecs = notarySpecs
-                ), externalTrace
-        ),
-        coerce = { it },
-        dsl = dsl,
-        initialiseSerialization = false
-)
+        dsl: RPCDriverDSL.() -> A
+): A {
+    return genericDriver(
+            driverDsl = RPCDriverDSL(
+                    DriverDSLImpl(
+                            portAllocation = portAllocation,
+                            debugPortAllocation = debugPortAllocation,
+                            systemProperties = systemProperties,
+                            driverDirectory = driverDirectory.toAbsolutePath(),
+                            useTestClock = useTestClock,
+                            isDebug = isDebug,
+                            startNodesInProcess = startNodesInProcess,
+                            waitForNodesToFinish = waitForNodesToFinish,
+                            extraCordappPackagesToScan = extraCordappPackagesToScan,
+                            notarySpecs = notarySpecs
+                    ), externalTrace
+            ),
+            coerce = { it },
+            dsl = dsl,
+            initialiseSerialization = false
+    )
+}
 
 private class SingleUserSecurityManager(val rpcUser: User) : ActiveMQSecurityManager3 {
     override fun validateUser(user: String?, password: String?) = isValid(user, password)
@@ -276,8 +144,8 @@ private class SingleUserSecurityManager(val rpcUser: User) : ActiveMQSecurityMan
 }
 
 data class RPCDriverDSL(
-        private val driverDSL: DriverDSL, private val externalTrace: Trace?
-) : DriverDSLInternalInterface by driverDSL, RPCDriverInternalDSLInterface {
+        private val driverDSL: DriverDSLImpl, private val externalTrace: Trace?
+) : InternalDriverDSL by driverDSL {
     private companion object {
         val notificationAddress = "notifications"
 
@@ -340,12 +208,20 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun <I : RPCOps> startInVmRpcServer(
-            rpcUser: User,
-            nodeLegalName: CordaX500Name,
-            maxFileSize: Int,
-            maxBufferedBytesPerClient: Long,
-            configuration: RPCServerConfiguration,
+    /**
+     * Starts an In-VM RPC server. Note that only a single one may be started.
+     *
+     * @param rpcUser The single user who can access the server through RPC, and their permissions.
+     * @param nodeLegalName The legal name of the node to check against to authenticate a super user.
+     * @param configuration The RPC server configuration.
+     * @param ops The server-side implementation of the RPC interface.
+     */
+    fun <I : RPCOps> startInVmRpcServer(
+            rpcUser: User = rpcTestUser,
+            nodeLegalName: CordaX500Name = fakeNodeLegalName,
+            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
+            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
+            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
             ops: I
     ): CordaFuture<RpcServerHandle> {
         return startInVmRpcBroker(rpcUser, maxFileSize, maxBufferedBytesPerClient).map { broker ->
@@ -353,7 +229,20 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun <I : RPCOps> startInVmRpcClient(rpcOpsClass: Class<I>, username: String, password: String, configuration: RPCClientConfiguration): CordaFuture<I> {
+    /**
+     * Starts an In-VM RPC client.
+     *
+     * @param rpcOpsClass The [Class] of the RPC interface.
+     * @param username The username to authenticate with.
+     * @param password The password to authenticate with.
+     * @param configuration The RPC client configuration.
+     */
+    fun <I : RPCOps> startInVmRpcClient(
+            rpcOpsClass: Class<I>,
+            username: String = rpcTestUser.username,
+            password: String = rpcTestUser.password,
+            configuration: RPCClientConfiguration = RPCClientConfiguration.default
+    ): CordaFuture<I> {
         return driverDSL.executorService.fork {
             val client = RPCClient<I>(inVmClientTransportConfiguration, configuration)
             val connection = client.start(rpcOpsClass, username, password, externalTrace)
@@ -364,7 +253,16 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun startInVmArtemisSession(username: String, password: String): ClientSession {
+    /**
+     * Starts an In-VM Artemis session connecting to the RPC server.
+     *
+     * @param username The username to authenticate with.
+     * @param password The password to authenticate with.
+     */
+    fun startInVmArtemisSession(
+            username: String = rpcTestUser.username,
+            password: String = rpcTestUser.password
+    ): ClientSession {
         val locator = ActiveMQClient.createServerLocatorWithoutHA(inVmClientTransportConfiguration)
         val sessionFactory = locator.createSessionFactory()
         val session = sessionFactory.createSession(username, password, false, true, true, locator.isPreAcknowledge, DEFAULT_ACK_BATCH_SIZE)
@@ -376,14 +274,23 @@ data class RPCDriverDSL(
         return session
     }
 
-    override fun <I : RPCOps> startRpcServer(
-            serverName: String,
-            rpcUser: User,
-            nodeLegalName: CordaX500Name,
-            maxFileSize: Int,
-            maxBufferedBytesPerClient: Long,
-            configuration: RPCServerConfiguration,
-            customPort: NetworkHostAndPort?,
+    /**
+     * Starts a Netty RPC server.
+     *
+     * @param serverName The name of the server, to be used for the folder created for Artemis files.
+     * @param rpcUser The single user who can access the server through RPC, and their permissions.
+     * @param nodeLegalName The legal name of the node to check against to authenticate a super user.
+     * @param configuration The RPC server configuration.
+     * @param ops The server-side implementation of the RPC interface.
+     */
+    fun <I : RPCOps> startRpcServer(
+            serverName: String = "driver-rpc-server-${random63BitValue()}",
+            rpcUser: User = rpcTestUser,
+            nodeLegalName: CordaX500Name = fakeNodeLegalName,
+            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
+            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
+            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
+            customPort: NetworkHostAndPort? = null,
             ops: I
     ): CordaFuture<RpcServerHandle> {
         return startRpcBroker(serverName, rpcUser, maxFileSize, maxBufferedBytesPerClient, customPort).map { broker ->
@@ -391,12 +298,21 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun <I : RPCOps> startRpcClient(
+    /**
+     * Starts a Netty RPC client.
+     *
+     * @param rpcOpsClass The [Class] of the RPC interface.
+     * @param rpcAddress The address of the RPC server to connect to.
+     * @param username The username to authenticate with.
+     * @param password The password to authenticate with.
+     * @param configuration The RPC client configuration.
+     */
+    fun <I : RPCOps> startRpcClient(
             rpcOpsClass: Class<I>,
             rpcAddress: NetworkHostAndPort,
-            username: String,
-            password: String,
-            configuration: RPCClientConfiguration
+            username: String = rpcTestUser.username,
+            password: String = rpcTestUser.password,
+            configuration: RPCClientConfiguration = RPCClientConfiguration.default
     ): CordaFuture<I> {
         return driverDSL.executorService.fork {
             val client = RPCClient<I>(ArtemisTcpTransport.tcpTransport(ConnectionDirection.Outbound(), rpcAddress, null), configuration)
@@ -408,13 +324,37 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun <I : RPCOps> startRandomRpcClient(rpcOpsClass: Class<I>, rpcAddress: NetworkHostAndPort, username: String, password: String): CordaFuture<Process> {
+    /**
+     * Starts a Netty RPC client in a new JVM process that calls random RPCs with random arguments.
+     *
+     * @param rpcOpsClass The [Class] of the RPC interface.
+     * @param rpcAddress The address of the RPC server to connect to.
+     * @param username The username to authenticate with.
+     * @param password The password to authenticate with.
+     */
+    fun <I : RPCOps> startRandomRpcClient(
+            rpcOpsClass: Class<I>,
+            rpcAddress: NetworkHostAndPort,
+            username: String = rpcTestUser.username,
+            password: String = rpcTestUser.password
+    ): CordaFuture<Process> {
         val process = ProcessUtilities.startJavaProcess<RandomRpcUser>(listOf(rpcOpsClass.name, rpcAddress.toString(), username, password))
         driverDSL.shutdownManager.registerProcessShutdown(process)
         return doneFuture(process)
     }
 
-    override fun startArtemisSession(rpcAddress: NetworkHostAndPort, username: String, password: String): ClientSession {
+    /**
+     * Starts a Netty Artemis session connecting to an RPC server.
+     *
+     * @param rpcAddress The address of the RPC server.
+     * @param username The username to authenticate with.
+     * @param password The password to authenticate with.
+     */
+    fun startArtemisSession(
+            rpcAddress: NetworkHostAndPort,
+            username: String = rpcTestUser.username,
+            password: String = rpcTestUser.password
+    ): ClientSession {
         val locator = ActiveMQClient.createServerLocatorWithoutHA(createNettyClientTransportConfiguration(rpcAddress))
         val sessionFactory = locator.createSessionFactory()
         val session = sessionFactory.createSession(username, password, false, true, true, false, DEFAULT_ACK_BATCH_SIZE)
@@ -427,12 +367,12 @@ data class RPCDriverDSL(
         return session
     }
 
-    override fun startRpcBroker(
-            serverName: String,
-            rpcUser: User,
-            maxFileSize: Int,
-            maxBufferedBytesPerClient: Long,
-            customPort: NetworkHostAndPort?
+    fun startRpcBroker(
+            serverName: String = "driver-rpc-server-${random63BitValue()}",
+            rpcUser: User = rpcTestUser,
+            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
+            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE,
+            customPort: NetworkHostAndPort? = null
     ): CordaFuture<RpcBrokerHandle> {
         val hostAndPort = customPort ?: driverDSL.portAllocation.nextHostAndPort()
         addressMustNotBeBound(driverDSL.executorService, hostAndPort)
@@ -452,7 +392,11 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun startInVmRpcBroker(rpcUser: User, maxFileSize: Int, maxBufferedBytesPerClient: Long): CordaFuture<RpcBrokerHandle> {
+    fun startInVmRpcBroker(
+            rpcUser: User = rpcTestUser,
+            maxFileSize: Int = ArtemisMessagingServer.MAX_FILE_SIZE,
+            maxBufferedBytesPerClient: Long = 10L * ArtemisMessagingServer.MAX_FILE_SIZE
+    ): CordaFuture<RpcBrokerHandle> {
         return driverDSL.executorService.fork {
             val artemisConfig = createInVmRpcServerArtemisConfig(maxFileSize, maxBufferedBytesPerClient)
             val server = EmbeddedActiveMQ()
@@ -471,10 +415,10 @@ data class RPCDriverDSL(
         }
     }
 
-    override fun <I : RPCOps> startRpcServerWithBrokerRunning(
-            rpcUser: User,
-            nodeLegalName: CordaX500Name,
-            configuration: RPCServerConfiguration,
+    fun <I : RPCOps> startRpcServerWithBrokerRunning(
+            rpcUser: User = rpcTestUser,
+            nodeLegalName: CordaX500Name = fakeNodeLegalName,
+            configuration: RPCServerConfiguration = RPCServerConfiguration.default,
             ops: I,
             brokerHandle: RpcBrokerHandle
     ): RpcServerHandle {
