+++ testing/node-driver/src/main/kotlin/net/corda/testing/internal/performance/Injectors.kt
@@ -3,7 +3,7 @@ package net.corda.testing.internal.performance
 import com.codahale.metrics.Gauge
 import com.codahale.metrics.MetricRegistry
 import com.google.common.base.Stopwatch
-import net.corda.testing.driver.ShutdownManager
+import net.corda.testing.internal.ShutdownManager
 import java.time.Duration
 import java.util.*
 import java.util.concurrent.CountDownLatch
